Add UseFloats option for comparing numbers
Adds a the `UseFloats` boolean field to `Options`. If true, then all
numberic values will be compared by first parsing them into float64
values and then comparing them using a scaled epsilon. This allows for
supressing diffs for numbers that may be the result of mathematic
operations, and thus suffer from a loss of precision. It also allows for
matching the behavior of Javascript and other runtimes that represent
all json numbers as 64 bit floats.